refactor where code is located
pnpm prepack still works as expected
