+++ src/runtime/composables/use-form.ts
@@ -1,56 +1,9 @@
 import type { z } from "zod"
-import { useFormKey } from "../../lib/core/composables/use-form-key"
-import { useFormStore } from "../../lib/core/composables/use-form-store"
-import { getComputedSchema } from "../../lib/core/utils/get-computed-schema"
+import { useAbstractForm } from "../../lib/core/composables/use-abstract-form"
 import type { UnwrapZodObject, UseFormConfigurationWithZod } from "../../types/types-adapters"
 import type { AbstractSchema, UseFormConfiguration } from "../../types/types-api"
 import type { DeepPartial, GenericForm } from "../../types/types-core"
 import { zodAdapter } from "../adapters/zod"
-
-function useAbstractForm<
-  Form extends GenericForm,
-  GetValueFormType extends GenericForm = Form,
->(
-  configuration: UseFormConfiguration<
-    Form,
-    GetValueFormType,
-    AbstractSchema<Form, GetValueFormType>,
-    DeepPartial<Form>
-  >,
-) {
-  const { schema } = configuration
-  const key = useFormKey(configuration.key)
-  const computedSchema = getComputedSchema(key, schema)
-  const initialStateResponse = computedSchema.getInitialState({
-    useDefaultSchemaValues: true,
-    constraints: configuration.initialState,
-    validationMode: configuration.validationMode ?? "lax",
-  })
-
-  const {
-    getHandleSubmitFactory,
-    getValidateFactory,
-    getValueFactory,
-    setValueFactory,
-    registerForm,
-    form,
-    formStore,
-  } = useFormStore<Form>(key)
-  registerForm(initialStateResponse.data)
-
-  const getValue = getValueFactory<Form, GetValueFormType>(form)
-  const setValue = setValueFactory(formStore, key, computedSchema)
-  const validate = getValidateFactory(form, key, computedSchema)
-  const handleSubmit = getHandleSubmitFactory(form, validate)
-
-  return {
-    handleSubmit,
-    getValue,
-    setValue,
-    validate,
-    key,
-  }
-}
 
 type UseFormReturnType<Form extends GenericForm, GetValueFormType extends GenericForm = Form> = ReturnType<typeof useAbstractForm<
   Form,
